+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+### 2.1.0, FIXME
+
+- Allow user to override User-Agent header
+- Change User-Agent format to report gem name, version and environment
+
 ### 2.0.0, 26.09.2017
 
 - There are **breaking** changes in this release, please read [upgrade notes](UPGRADE_NOTES.md#v1---v2)